refactor(layout): PageHeader 컴포넌트화 및 폴더 구조 개선 [#43]
- PageHeader 컴포넌트 구현 및 전체 페이지 적용
- components 폴더 구조 도메인별 재정리
