+++ components/profile/WithdrawalConfirmModal.tsx
@@ -1,13 +1,13 @@
 'use client';
 
 import { withdrawUser } from '@/app/(tabs)/users/[username]/edit/actions';
-import FormButton from '@/components/ui/form/FormButton';
-import FormInput from '@/components/ui/form/FormInput';
 import { ACCOUNT_VALIDATION } from '@/lib/constants';
 import { useRouter } from 'next/navigation';
 import { startTransition, useActionState, useEffect } from 'react';
 import { createPortal } from 'react-dom';
 import { toast } from 'sonner';
+import FormButton from '../form/FormButton';
+import FormInput from '../form/FormInput';
 
 interface IWithdrawalConfirmModalProps {
   onCloseAction: () => void;
