+++ components/response/ResponseViewMode.tsx
